pyproject.toml
[project]
name = "ipig-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "delightening", email = "jason4617987@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
